+++ packages/backtest/src/core/index.ts
@@ -1,2 +1,0 @@
-export { default as Broker } from './Broker'
-export { default as Strategy } from './Strategy'