+++ packages/trade/src/index.ts
@@ -2,6 +2,7 @@ import createJsEnv from './trading/env'
 import helpers from './helpers'
 import * as indicators from './indicators'
 
+export * from './types'
 export {
   helpers,
   indicators,